mfd_upd350: new multi-function device driver for Microchip UPD350.
The UPD350 is marketed as a "Highly Integrated Small Form Factor
USB Type-C® Power Delivery 3.0 Port Controller" but in particular it
provides both the PD MAC and 10 GPIO lines that are independent from the
PD interface.

While the chip is available in both I2C and SPI variants, this commit
only implements the SPI interface (which is what UPD301C uses), but
leaves enough space to implement the I2C interface later on, following
the shape of the gpio_mcp23sxxx driver.

Signed-off-by: Diego Elio Pettenò <flameeyes@meta.com>