usbc_vbus: make all APIs besides check_level optional.
Even in the usb_c subsys samples, the only API that is always called is
check_level. Everything else appears closer to a quirk of how it is
implemented with the ADC driver.

For UPD350, these other APIs cannot be implemented, as it includes its
own comparator. Instead of returning -ENOTSUP on the driver itself,
make it generic enough as other controllers are likely lacking the
ability to take a raw measurement as well.