doc: move puncover down on the page.
The puncover tool does not appear to be maintained anymore and it is
affected by multiple issues. Providing it as the first in the list
is a disservice to users as they might get frustrated and not notice
the `ram_report` tool that actually works.

Move the information at the bottom of the section, and add a warning
about the fact that this is a third party tool that might or might
not be working.

(Also make the link consistent to avoid "here" being a hyperlink, it
adds nothing to the reader either in RST or HTML.)

Signed-off-by-: Diego Elio Pettenò <flameeyes@meta.com>